Move trajectory analysis helper modules to analib/ subfolder
