+++ src/fairmd/lipids/analib/maicos.py
